+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* More complete documentation for the `WebArchive` class.
 ### Fixed
 * Raise a `WebArchiveError` when attempting to extract a webarchive with no main resource.
 ### Changed